# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Schedurx Backend API
  version: 1.0.0
  description: >
    Clinic business API — patients, doctors, appointments, and calendar tools.
    The /tools/* paths are called directly by the Ultravox voice agent during a
    live call; /internal/* is called only by schedurx-ultravox-demo-api.
servers:
  - url: https://backend.example.test
    description: Replace with this service's public URL for deployed environments.
components:
  securitySchemes:
    ToolsBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: TOOLS_API_KEY — configured into Ultravox's tool definitions by the demo bridge.
    InternalBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: INTERNAL_API_KEY — held only by schedurx-ultravox-demo-api.
    FirebaseBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        A Firebase ID token for a Staff account. Custom claims (role, clinicId,
        doctorId) are set at onboarding time via POST /internal/staff and read
        by every /api/v1/* route — clinicId is always sourced from these claims,
        never from the request itself.
    NettuWebhookKeyAuth:
      type: apiKey
      in: header
      name: nettu-scheduler-webhook-key
      description: >
        Static key set once via `PUT /api/v1/account/webhook` on the nettu
        account (see scripts/setup-nettu-webhook.js) — nettu echoes this exact
        value back on every reminder callback. Not a bearer token.
    TwilioSignatureAuth:
      type: apiKey
      in: header
      name: X-Twilio-Signature
      description: >
        HMAC-SHA1 signature over the exact request URL + form params, verified
        against TWILIO_AUTH_TOKEN (see twilio-client.js's validateSignature).
        Twilio computes and sends this automatically; it isn't a credential a
        caller sets manually.
  schemas:
...
    StructuredToolError:
      type: object
      required: [success, error]
      properties:
        success:
          const: false
        error:
          type: object
          required: [code, message, details]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type:
                - object
                - "null"
              additionalProperties: true
...
    ApiV1Success:
      type: object
      required: [success, data, message]
      properties:
        success:
          const: true
        data:
          type: object
          additionalProperties: true
        message:
          type:
            - string
            - "null"
    ApiV1Error:
      allOf:
        - $ref: "#/components/schemas/StructuredToolError"
...
    Visit:
      type: object
      properties:
        id:
          type: string
        clinicId:
          type: string
        patientId:
          type: string
        doctorId:
          type: [string, "null"]
        appointmentId:
          type: [string, "null"]
        visitDate:
          type: string
          format: date
        symptoms:
          type: [string, "null"]
        notes:
          type: [string, "null"]
        followUpDate:
          type: [string, "null"]
          format: date
...
paths:
...
  /api/v1/visits:
    get:
      summary: List visits for a patient.
      security: [{ FirebaseBearerAuth: [] }]
      parameters:
        - { name: patientId, in: query, required: true, schema: { type: string } }
      responses:
        "200":
          description: Visit history for the patient.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiV1Success" }
    post:
      summary: Create a visit record.
      security: [{ FirebaseBearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Visit" }
      responses:
        "201":
          description: Visit created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiV1Success" }
  /api/v1/visits/{id}:
    patch:
      summary: Update a visit (notes, follow-up date, Rx flags).
      security: [{ FirebaseBearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Updated visit.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiV1Success" }
...
  /api/v1/threads:
    get:
      summary: List consult threads for the caller's clinic.
      security: [{ FirebaseBearerAuth: [] }]
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [open, escalated, closed] } }
      responses:
        "200":
          description: Threads.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiV1Success" }
  /api/v1/threads/{id}/messages:
    get:
      summary: List messages in a thread.
      security: [{ FirebaseBearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Messages, oldest first.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiV1Success" }
    post:
      summary: >
        Send a reply (doctor/owner only). Persists regardless of WhatsApp
        delivery status — see src/services/whatsapp-client.js, currently stubbed.
      security: [{ FirebaseBearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties: { body: { type: string } }
      responses:
        "201":
          description: Message sent (or queued, if WhatsApp isn't configured yet).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiV1Success" }
        "403":
          description: Caller's role is not doctor/owner.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiV1Error" }
  /api/v1/threads/{id}/escalate:
    post:
      summary: Escalate a thread to critical (receptionist/owner only).
      security: [{ FirebaseBearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Updated thread.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiV1Success" }
        "403":
          description: Caller's role is not receptionist/owner.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiV1Error" }
  /api/v1/threads/{id}/read:
    patch:
      summary: Clear the unread count on a thread.
      security: [{ FirebaseBearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Updated thread.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiV1Success" }
...
  /api/v1/visits/{id}/upload-url:
    post:
      summary: Get a Supabase Storage signed upload URL for an attachment.
      description: The browser PUTs the file bytes directly to Storage — never through this server.
      security: [{ FirebaseBearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fileName]
              properties:
                fileName: { type: string }
                contentType: { type: [string, "null"] }
      responses:
        "200":
          description: "{ path, uploadUrl }"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiV1Success" }
  /api/v1/visits/{id}/attachments:
    post:
      summary: Record a completed direct-to-storage upload on the visit.
      security: [{ FirebaseBearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [path, type]
              properties:
                path: { type: string }
                type: { type: string, enum: [photo, digital] }
      responses:
        "201":
          description: Attachment recorded.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiV1Success" }
  /api/v1/visits/{id}/attachments/read-url:
    get:
      summary: Short-lived signed read URL for a private-bucket attachment.
      security: [{ FirebaseBearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - name: path
          in: query
          required: true
          description: Must be prefixed "{clinicId}/{visitId}/" — otherwise 403.
          schema: { type: string }
      responses:
        "200":
          description: "{ url }"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiV1Success" }
        "403":
          description: path does not belong to this clinic/visit.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiV1Error" }
  /api/v1/visits/{id}/recap:
    post:
      summary: Turn a doctor's recap (typed or recorded) into a structured clinical note, saved onto the visit.
      description: Backs both manual "Recap" capture mode and ambient capture.
      security: [{ FirebaseBearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text: { type: [string, "null"] }
                audioBase64: { type: [string, "null"] }
                filename: { type: [string, "null"] }
      responses:
        "200":
          description: "{ visit, transcript }"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiV1Success" }
        "422":
          description: Neither text nor audioBase64 given, or the recording was silent.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiV1Error" }
        "503":
          description: OPENAI_API_KEY is not configured.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiV1Error" }

  # ── /api/v1/call-logs, /api/v1/wa-logs ────────────────────────────────────────
...
  /api/v1/notifications/{id}/read:
    patch:
      summary: Mark one notification read.
      security: [{ FirebaseBearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Updated notification.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiV1Success" }
...